requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
gcloud==0.18.3
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
jws==0.1.3
oauth2client==4.1.3
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.4
pyparsing==2.4.7
Pyrebase4==4.5.0
pyTelegramBotAPI==4.1.0
python-dotenv==0.19.0
python-jwt==2.0.1
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
rutimeparser==1.1.1
schedule==1.1.0
six==1.16.0
tzdata==2021.2.post0
urllib3==1.26.7
twilio==7.4.0
